requirements.txt
Flask==2.0.2
Flask-WTF==1.0.0
bcrypt==3.2.0
SQLAlchemy==1.4.23
flask_sqlalchemy
yahoofinancials
simplejson
yahooquery
flask_wtf
flask_bcrypt
Authlib
python-dotenv
requests
auth0-python